Split context and provider, fix an example in README
