+++ src/index.tsx
@@ -1,4 +1,4 @@
-import { PubNubProvider } from './context/PubNubContext';
+import { PubNubProvider } from './context/PubNubProvider';
 import { PubNubConsumer } from './context/PubNubConsumer';
 import { usePubNub } from './hooks/usePubNub';
